Removed unnecessary topics for CitAtlas and moved streets data into districts.
