--- src/data-sources/election-next.js
@@ -1,10 +1,0 @@
-export default {
-  id: 'nextElectionAPI',
-  type: 'http-get',
-  url: 'https://apis.philadelphiavotes.com/election',
-  options: {
-  },
-  success: function(data) {
-    return data;
-  }
-}
