--- src/data-sources/police-distr.js
@@ -1,8 +1,0 @@
-export default {
-  id: 'policeDistr',
-  url: 'https://services.arcgis.com/fLeGjb7u4uXqeF9q/ArcGIS/rest/services/Police_District/FeatureServer/0',
-  type: 'esri',
-  options: {
-    relationship: 'contains',
-  },
-};
