--- src/data-sources/police-psa.js
@@ -1,8 +1,0 @@
-export default {
-  id: 'policePSA',
-  url: 'https://services.arcgis.com/fLeGjb7u4uXqeF9q/ArcGIS/rest/services/Boundaries_PSA/FeatureServer/0',
-  type: 'esri',
-  options: {
-    relationship: 'contains',
-  },
-};
