+++ src/main.js
@@ -26,25 +26,17 @@ import greeting from './general/greeting';
 import threeOneOne from './data-sources/311';
 import threeOneOneCarto from './data-sources/311-carto';
 import assessmentHist from './data-sources/assessment-history';
-import charterSchools from './data-sources/charter-schools';
 import condoList from './data-sources/condo-list';
 import crimeIncidents from './data-sources/crime-incidents';
-import divisions from './data-sources/divisions';
 import dorCondoList from './data-sources/dor-condo-list';
 import dorDocuments from './data-sources/dor-documents';
-import electedOfficials from './data-sources/elected-officials'
-import fireStation from './data-sources/fire-station';
 import liBusinessLicenses from './data-sources/li-business-licenses';
 import liInspections from './data-sources/li-inspections';
 import liPermits from './data-sources/li-permits';
 import liPermitsAdditional from './data-sources/li-permits-additional';
 import liViolations from './data-sources/li-violations';
 import nearbyZoningAppeals from './data-sources/nearby-zoning-appeals';
-import nextElectionAPI from './data-sources/election-next';
 import opa from './data-sources/opa';
-import pollingPlaces from './data-sources/polling-places';
-import policeDistr from './data-sources/police-distr';
-import policePSA from './data-sources/police-psa';
 import rco from './data-sources/rco';
 import regmaps from './data-sources/regmaps';
 import sanitationCenters from './data-sources/sanitation-centers';
@@ -68,11 +60,6 @@ import threeOneOneTopic from './topics/311';
 import stormwaterTopic from './topics/stormwater';
 import nearby from './topics/nearby';
 import districts from './topics/districts';
-import safety from './topics/public-safety';
-import schools from './topics/schools';
-import streets from './topics/streets';
-import trash from './topics/trash';
-import voting from './topics/voting';
 
 // styles
 // TODO move all styles here (that have a npm package)
@@ -138,25 +125,17 @@ mapboard({
     threeOneOne,
     threeOneOneCarto,
     assessmentHist,
-    charterSchools,
     condoList,
     crimeIncidents,
-    divisions,
     dorCondoList,
     dorDocuments,
-    electedOfficials,
-    fireStation,
     liBusinessLicenses,
     liInspections,
     liPermits,
     liPermitsAdditional,
     liViolations,
     nearbyZoningAppeals,
-    nextElectionAPI,
     opa,
-    policeDistr,
-    policePSA,
-    pollingPlaces,
     rco,
     regmaps,
     sanitationCenters,
@@ -179,11 +158,6 @@ mapboard({
     threeOneOneTopic,
     stormwaterTopic,
     nearby,
-    safety,
-    schools,
-    voting,
-    trash,
-    streets,
     districts,
   ],
   components: [
