--- src/topics/trash.js
@@ -1,88 +1,0 @@
-import transforms from '../general/transforms';
-const getNearest = transforms.getNearest.transform;
-
-
-export default {
-  key: 'trashDay',
-  icon: 'trash',
-  label: 'Trash & Recycling',
-  dataSources: [
-    'sanitationCenters',
-  ],
-  components: [
-    // {
-    //   type: 'ExternalLink',
-    //   options: {
-    //     data: function(state) {
-    //       return state.geocode.data.properties.opa_account_num;
-    //     },
-    //     href: function(state) {
-    //       var id = state.geocode.data.properties.opa_account_num;
-    //       return '//property.phila.gov/?p=' + id;
-    //     }
-    //   }
-    // },
-    {
-      type: 'badge',
-      options: {
-        titleBackground: '#58c04d',
-        // externalLink: {
-        //   data: function(state) {
-        //     return 'the opa number is ' + state.geocode.data.properties.opa_account_num + ' got it?';
-        //   },
-        //   href: function(state) {
-        //     var id = state.geocode.data.properties.opa_account_num;
-        //     return '//property.phila.gov/?p=' + id;
-        //   }
-        // }
-      },
-      slots: {
-        title: 'Your Trash Day Is',
-        value: function(state) {
-          return transforms.dayofweek.transform(state.geocode.data.properties.rubbish_recycle_day);
-        },
-      },
-    },
-    {
-      type: 'vertical-table',
-      options: {
-        nullValue: 'None',
-        externalLink: {
-          action: function() {
-            return 'Report missed trash collection, illegal dumping, or other issues';
-          },
-          href: function(state) {
-            return '//www.philadelphiastreets.com/helpful-links/request-a-service';
-          }
-        }
-      },
-      slots: {
-        title: 'Trash and Recycling',
-        fields: [
-          {
-            label: 'Nearest Sanitation Convenience Center',
-            value: function(state) {
-              // REVIEW getNearest function is called before property is ready, giving a null error
-              return (`${getNearest(state, 'sanitationCenters', 'distance').name} <br>\
-              ${getNearest(state, 'sanitationCenters', 'distance').address} <br>\
-              ${getNearest(state, 'sanitationCenters', 'distance').phone} <br>
-              All locations are open Mon. through Sat. \
-              between 8 a.m. and 8 p.m. Except on \
-              City holidays.<br>\
-              <a href="//www.philadelphiastreets.com/binrequest/" target="_blank">Get a recycling bin. </a>\
-              `);
-            },
-          },
-          {
-            label: 'Leaf Collection',
-            value: function(state) {
-              return state.geocode.data.properties.leaf_collection_area;
-            },
-          },
-        ]
-      }
-    }, // end streets table
-  ],
-  identifyFeature: 'address-marker',
-  parcels: 'pwd'
-}
